release 2.1.0 and add documentation readme
* readme now explains in detail a number of use cases and describes the
various enhancements supported and the internal methods available for
reuse.
* enhancements to 2.1.0 and the package is built
+ 2.1.0: * add 'autoaync.refreshEditable' method - reloads items in an
".inline-editable-host" or an item repeater on demand.  The url of the
host removes any parameters in the url for that get/post that are
defined as data attributes on the button.
* add 'autoasync.urlRemoveParams' method (used by refreshEditable) .
autoasync.clickButton added an automatic complete callback to
refreshEditable on any element selectors defined in a
'data-host-selector' attribute of the button.
* 'inline-editable-button' elements can now be anywhere within an
enhanced element. It can specify an 'data-editable-container' as an
alternate to being in an '.inline-editable-host"