Many enhancements and new plugins
* Now the document can only enhanced once to prevent multiple click
event handlers on entities.  However, duplicate calls to init can be
made to add additional callbacks when a section is enhanced.
* enhance now accepts an additional enhanceType parameter used by some
plugins.  The repeater is one such plugin that passes whether the list
or an item is being enhanced.
* refreshing an editable from a JSON AJAX response now handles a null
response
*  a null response from an inline-editable-host AJAX response will now
trigger insertion of a noitems template, if a data-template-noitems
attribute exists with a template name.
* isValid base method handles more edge cases and is renamed to indicate
more than a form can be validated.
* a button that triggers a form validation normally can now be decorated
with a 'novalidate' class so that validation does not occur.  This is
for instance to be used for cancel buttons.  The 'link-button-cancel'
class will be deprecated and removed in a future release.  This change
is to be more in line with standards.
* AJAX post responses will now check for a redirect parameter on the
response, and redirect to the url provided.
* AJAX post response will now check for a cookie result parameter, and
if the cookie plugin is available it will add those cookie values to the
page cookie for the next post.
* AJAX post response now checks for a "reload" parameter and will reload
the page if set.
* AJAX post response will now check for multiple items being returned
from the 1 posted.  If so, those items are appended with the configured
template for that editable host.  This allows saving one item in a post
to result in multiple result items.
* AJAX responses can now return the JSON items in more ways.  It may be
an Items property, items property, or the main object can be an array of
items or a single item.
* datarepeater plugin is enhanced to allow the returned list of items to
update multiple repeaters from that single AJAX call.  This reduces
duplicate calls to the server for the same data in different
presentations on the page.  Added to support the carousel that needs an
photo for each slide and another list with the buttons to switch to each
slide.
* New Plugins: listbuildersource, changeupdate, enhancelist, and
carousel (for boostrapper carousels).  Documentation for each to come
later.
* Added wizard capability via  separate file
* added external reference for prettyDate plugin
* all changes thoroughly tested in a production site for 6+ months